Move state machine things, caught bad removals
